// src/app/models/bank-accounts.model.ts
export class BankAccountModel {
    id: string;
    mobileNo: string;
    accNum: string;
    accType: string;
    accIfsc: string;
    accName: string;
    bankName: string;
    bankCode: string;
    bankBranch: string;
    debitCardNo: string;
    debitCardExpiry: string;
    debitCardPin: string;
    currentOtp: string;
    netAmount: string;
}